Change types to work nicely with MSVC.
ssize_t is not MSVC supported type, and we do not need more than 31 bits
of precision, so replace with size_t.
Many places we were passing size_t for resource counts which ended up being
cast to uint32_t. Make these consistently uint32_t.
Typedef std::vector<std::pair<ResourceId, uint32_t>> to ResourceList, as
this is used everywhere.

+++ cc/replayd/src/Connection.cpp
@@ -22,7 +22,7 @@ namespace android {
 namespace caze {
 
 bool Connection::sendString(const std::string& s) {
-    uint32_t length = s.size();
+    uint32_t length = static_cast<uint32_t>(s.size());
     if (this->send(&length, sizeof(length)) != sizeof(length)) {
         return false;
     }
